Drop SDL frontend entirely
This was a long way coming. It was initially created as a way of
running FreeJ2ME on the RaspBerry Pi, but once the libretro
frontend became more usable, there wasn't much of a need for it,
so it was then repurposed as a way of integrating FreeJ2ME-Plus
with libTAS... and as promising as that looked back then, the Java
VM has limitations that completely break the deal of doing any kind
of actual TAS (no serialization, can't rewind, no save states, etc)

So after laying around in serious state of disrepair, i think it's
time to pull the plug on this. It requires a completely new UI
interface to become anywhere close to how useful AWT is, and it
has may inherent issues due to how it works in order to integrate
with libTAS, like requiring inputs reads be tied to new frames, when
some apps wait for inputs before pushing new frames.

A sincere thank you for those that contributed to this frontend up
to this point.