PlatformGraphics: Use drawRGB for (almost) everything image related
Since drawRGB is at a very mature stage, it should be able to handle
all calls that were resolved into a awt.Graphics.drawImage() without
scaling.

Overall performance is about the same (sometimes faster than drawImage,
sometimes slower), but the boon here is having less of a dependence
on the awt packages, and having drawRGB receive a lot more attention
as it'll be the backbone for almost all image operations.

This commit also improves the graphics object's usage consistency,
as some outer LCDUI classes like TextField, Gauge, etc were using
graphics.getGraphics2D().* methods, which would cause the translation
variables to be inconsistent with the graphics object's actual state,
since they wouldn't be updated due to the translation being done
directly to awt. This was also the case for M3G's Graphics3D.

And going into more detail in M3G, it's untextured polygon rendering
will regress with this commit, as there will be no interpolation
between vertex colors. This is expected, and the previous code wasn't
correct either... this is an effort to also improve its graphics
object usage consistency and in turn help future patches on it.