RecordStore: Use UTF-8 only if mobile encoding is ISO_8859-1
Fixing a regression introduced in the last commit, and improving
on it at the same time. Helps MIDP saves that contain chinese and
special non-ISO_8859-1 characters. If the current encoding is
anything other than ISO_8859-1, RecordStore will use said encoding
for save folder names (Shift_JIS, EUC_KR) which i think is the
correct way to do this.