M3G: Improve how object references are handled.
The previous system for those was not correct, and more complicated
than it should be, with any relevant Objects requiring an override of
the base Object3D's findID and doGetReferences methods. Now this is
no longer the case and the whole implementation is more in line with
the M3G spec. Fixes a few things like the tornado in Tornado Mania
finally being on screen for example.

Duplication logic still has to be reworked though.