M3G: Clean up and optimize rendering path.
Quite a few changes with this one, most notable are:

- Untextured/Wireframe mode now share the same code path as textured
triangles, making Graphics3D's main render() method much smaller. Also
restores the previously regressed untextured rendering mode... although
it still isn't perfect at this stage.

- The 'Triangle' class is now the one responsible for culling and
clipping triangles instead of Graphics3D's render() method, which
again makes said method smaller, and allows the Triangle class
to do some additional optimizations to the whole process, like not
needing a second look at the whole triangle array in order to cull
and clip triangles, reusing the same index buffer array for all triangles
while only using the triangle index to access the expected indices, and also
only allocating the texture coordinates into the triangle if said triangle
is actually going to be rendered. (can probably be optimized further to not
even allocate new vertex positions for the triangle until it's deemed
renderable, but would require a large rework of the entire class and
ancillary methods for transformations).

- Fog blending now works in an actual use case (Tornado Mania 3D), albeit
with a hack that multiplies the final fog factor by a magnitude of 250 in
order to approximate how it should actually look. There is probably an
issue with how the triangles' Z coordinate is being resolved, as the
vast majority of cases has them going from 0.99 to 1.00, which is way
too tight of an interval.

- Due to the reworked render path, Graphics3D's render() method no
longer relies on awt Graphics' methods for rendering, which should
be a plus for portability.

- Some other improvements were made here and there to reuse variables
and simplify a few calculations.