AudioClip: More off-spec improvements (MMF/SMAF of course)
In usual SMAF-On-Samsung-AudioClip fashion, specification be damned.

Some versions of Snowball Fight pass whatever they want as the clip
type when creating an AudioClip. This, of course, would cause issues
whenever said value wasn't in the expected [1-3] interval. AudioClip
specification dictates that an IllegalArgumentException should be
thrown whenever that happens.

The end result is that this check has to be dropped, and in its place
is now a force-cast to TYPE_MMF, as it's always the problem child in
AudioClip.

MIDI looping was also improved.