RecordStore: Fix setRecord() implementation.
We should always allocate a byte array of 'numBytes' size to then
set the incoming data into. Previously it reused the currently saved
array on the requested set position, which could be larger or smaller
than the new data.

Fixes CAPCOM's "Don't Pull" and likely any others that rely on the
proper setRecord behavior.