EXPERIMENTAL: PlatformPlayer: Rework volume and panning changes again
Now volume changes on MIDI-based players are set through the Master Volume
SysEx message. It actually works now because the synths are guaranteed
to be there before the players are even realized. Fixes a bunch of
issues regarding volume setup while also simplifying the whole process
a lot.

Note that this still caused some smaller regressions though, as is standard
with the Java Sound API. Most notable being WALL-E's song fadein/fadeout
now crackling a bit, but it's a sacrifice that i'm more than willing to make
in exchange for FAR more reliable and simple volume control.

Related to #145