Implement support for DoJa and KDDI's tempo changes.
We have a whole "tempoControl" class in PlatformPlayer that was
meant to be used by MIDP stuff, and nothing there really seems to
use it, so now we're extending it all the way to DoJa and KDDI as
well, where some apps actually set different tempo values.

This commit also makes some minor improvements to volume changes.

Related to #145