Don't close FreeJ2ME-Plus on app-requested close, and other stuff
Now the user gets a choice on whether to close FreeJ2ME, or load
a new application on it.

Beyond that, GameCanvas only renders to screen if it is visible,
fixes some other versions of Shinobi 2.

Also, PAC-MAN for S40v1 tries passing OTT files as WAV files in
Nokia Sound API... it's the first game to ever do this, so i'll
assume that for all cases where the header isn't a valid MIDI or WAV,
it's actually an OTT.